+++ frigate/data_processing/real_time/face.py
@@ -20,7 +20,7 @@ from frigate.comms.event_metadata_updater import (
 from frigate.comms.inter_process import InterProcessRequestor
 from frigate.config import FrigateConfig
 from frigate.const import FACE_DIR, MODEL_CACHE_DIR
-from frigate.data_processing.common.face.model import (
+from frigate.data_processing.common.face.recognizer import (
     ArcFaceRecognizer,
     FaceNetRecognizer,
     FaceRecognizer,
